+++ CHANGELOG.md
@@ -26,6 +26,7 @@ By default only inline images are converted because other layouts do not have a 
 - Preserve page borders in RTF ➡️ DOCX converter
 - Improve logic for VML pictures and wrap layouts and in DOCX ➡️ RTF converter
 - Support for VML shapes and text boxes in DOCX ➡️ RTF converter
+- Fix: the rotation angle of shapes was not accurate in DOCX ➡️ RTF converter
 - Fix various minor bugs
 - Code cleanup
 - Update dependencies